Item1904: Move IE-specific CSS to the conditional section in css.pattern.tmpl, because that IE-specific CSS makes Firefox truncate pre blocks when inside TinyMCE.
I cannot see what difference that IE-specific CSS makes; it appears to make no difference to IE 8 in any of its various "compatibility" modes. I suppose it is there for older versions of IE. I can see the "1%" CSS when I view source, so I assume it will work where it is needed.


git-svn-id: http://svn.foswiki.org/trunk@4622 0b4bb1d4-4e5a-0410-9cc4-b2b747904278