Item2369: Add rowspan support to the WYSIWYG editor This also fixed a bug in how Web.WikiWord links are rendered, if the link is the only text in the table cell.
Tables are no longer "normalised" when converting from HTML to TML by appending |'s so that every row has the same number of |'s. The logic to do so when supporting rowspans is complex (read: likely to be buggy) and the "normalisation" feature is seldom used (it only had any effect if the HTML has a table with a hole in it.) Taking out the normalisation does not result in loss of content. 


git-svn-id: http://svn.foswiki.org/trunk@5573 0b4bb1d4-4e5a-0410-9cc4-b2b747904278