Item8706: configure dies if either $Foswiki::cfg{HttpCompress} or $Foswiki::cfg{Cache}{Compress} is not defined. Maybe it dies because I have set FOSWIKI_ASSERTS, but this would otherwise emit a warning into the apache error log, which isn't good either.
This change will also deal with the (odd and perhaps unlikely) condition where the two settings have different values, both of which are true. In that case, the warning would be reported, even though the settings are consistent. This might happen if the admin modifies one or the other in the .cfg file and not via configure.


git-svn-id: http://svn.foswiki.org/trunk@7424 0b4bb1d4-4e5a-0410-9cc4-b2b747904278