Item9948: registerMETA API changes. Crawford, Sven, please check...
I'm not *really* sure that changes near line 61 in Query/Node.pm are
appropriate. The problem is that we used to clear the %isArrayType hash
when the .pm is compiled. So any plugin that tries to add a new element,
must wait til the 2nd request under fastcgi/fcgid/mod_perl, and I assume
under plain old CGI it's actually impossible.

So I added a param to registerMETA. This was broken anyway, so I assume nobody
was using it.

Anyway, this fix allows me to register a new META:FOO that can be treated as an
array type like META:FILEATTACHMENT. Without this fix, my multi-value META:FOO
is only able to return the first value in a QUERY on a specific topic.

git-svn-id: http://svn.foswiki.org/trunk@9848 0b4bb1d4-4e5a-0410-9cc4-b2b747904278