Item10097: Add unit tests (fail). Cause is supporting protected USERSWEB, Item9809
Sorry, I struggled to make the tests fail inside TopicUserMappingTests.pm, so created
a new TopicUserMappingAsGuestTests.pm instead.

Commenting-out line 1488 in Foswiki::Users::TopicUserMapping.pm makes the test pass.

The problem is that we temporarily change the {session}->{user} = admin, but then
those %USERSWEB%.*Group topics are primed into the Foswiki::MetaCache when the
'actual' user may not necessarily have VIEW access.

Effectively this means that Foswiki::search only checks ACLs on topics once: the
first time it scans over it in a search. If that first time occurs while we happen
to be in TopicUserMapping when {session}->{user} = admin, then the Foswiki::MetaCache
may be contaminated with *Group topics that were supposed to be protected from the
actual user.

git-svn-id: http://svn.foswiki.org/trunk@10270 0b4bb1d4-4e5a-0410-9cc4-b2b747904278