Item10284: Generate alt. regex for META items
This fix adds a new option - inMeta - for Render::getReferenceRE.
When a non-WikiWord topic is renamed, the META RE is changed to be an
exact match without delimiters.

The unit tests still pass, but this is fragile, so I'd like to get some
feedback before committing to the Release branch.
:

git-svn-id: http://svn.foswiki.org/branches/Release01x01@10869 0b4bb1d4-4e5a-0410-9cc4-b2b747904278