Item10905: Permit spaces in link query string
Add unit tests for legacy mailto links with and without query strings.

Fix auto test that expected ?query strings to work in legacy links.

Add {AntiSpam}{EntityEncode} with default enabled.  Email addresses
will have all non-alphanumeric characters encoded to HTML hex entities.

Revert the previous fix to this task that removed legacy support for
[[mailto Link Text]] style links.

Add exception for recognizing space delimited square bracket links. This
format was deprecated in 2005.  If a ? is detected in the URL part of
the legacy format,  then any spaces are assumed to be part of the query
string.

[[mailto:you@here.com?subject=test subject]]   Before this code, the
mailto link would be generated with link text of "subject".  After this
commit, subject becomes part of the query string and the entire url
becomes the link text.

git-svn-id: http://svn.foswiki.org/trunk@12040 0b4bb1d4-4e5a-0410-9cc4-b2b747904278