Item11183: Add the LISTNONEMPTY function
In the unit test, discovered a subtle issue with LEFT (or any function
that retrieves a range of cells into a list.)  In Item9400, cells
containing a list are flattened into the list.  But the fix is in the
wrong place.  The extra split is needed on the results of _getTableRange,
not on an item that was alreay split.

Also when using a range function like LEFT, an empty cell is added
into the internal list as a single space, rather than an empty entry.
The twiki code to remove empty entries was not detecting the space as
empty.

git-svn-id: http://svn.foswiki.org/trunk@12778 0b4bb1d4-4e5a-0410-9cc4-b2b747904278