Item11195: Dates are decrementing when edited
The issue happens in timezones west of GMT, which have a negative
offset, and when Foswiki is also configured to display dates in
servertime instead of gmtime.

The Fosiwki Time parse routine assumes GMT unless the defaultLocal flag
is set.  Changing the render for edit code to enable localtime if the
server is configured to display dates using servertime.

This needs someone with positive timezone offsets to test using gmtime
and servertime to make sure it doesn't break anything for them.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@13076 0b4bb1d4-4e5a-0410-9cc4-b2b747904278