Item11386: Add a sanity check for the Query settings
Note:  This does not need to be pulled into Foswiki trunk - 1.2 uses an
alternate way of detecting the query operators.  This issue can probably
only happen if a 1.0 system has been upgraded to 1.1.

Also, a checker won't work here because the setting is Hidden, and the
checkers are not executed for hidden settings.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@13512 0b4bb1d4-4e5a-0410-9cc4-b2b747904278