Item11409: Flush password cache, move lockfile
Added expert config parameter to verify password file timestamp is not
newer than the cached copy and clear the cached copy if it's out of
date.  Also renamed the lock file to be the htpasswd filename with .lock
appended.

These changes should permit two Foswiki installs to share a common
password file.

Also updated the advice on choosing a password hash type.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@13555 0b4bb1d4-4e5a-0410-9cc4-b2b747904278