Item11441: redirectto overrides endPoint
The old CommentPlugin uses the query parameter "redirectto" to return to
a different topic after save.   The REST script doesn't implement this,
and instead redirects to query parameter "endPoint".

Change the comment rest handler to replace endPoint with the value
recovered from redirectto and print a warning.  Add unit tests to
verify that the redirect can handle anchors and query parameters, and
that a old UserCommentTemplate with redirectto correctly redirects.

git-svn-id: http://svn.foswiki.org/trunk@13648 0b4bb1d4-4e5a-0410-9cc4-b2b747904278