Item11450: Spreadsheet Plugin ABOVE references r0
It was hard-coded to R0.  In looking through the tests, also
noticed that LEFT was hard-coded to C0.  The rest of the plugin seems to
calculate correctly.

Added a test for the actual values returned from LEFT,  and correct the
ABOVE test to expect R1, not R0.   Added test for SUMPRODUCT and a
combined SET / GET test both of which were initially suspected as
causing the issue.

Update copyright for 2012.  Bump version.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@13683 0b4bb1d4-4e5a-0410-9cc4-b2b747904278