Item652: Item710:  Item2196: Failed registration
These tasks all report the failure where the misconfigured email
server leaves partially registered users that require manual intervention.

I don't see how this prevents spam registration.  If Net::SMTP or
sendmail is configured correctly, bad domains or user addresses
result in a bounce, not a SMTP protocol failure.  At least I've not been
able to create an issue where sending email fails with a correctly
configured server.

This fix removes the "backout" of the user registration when SMTP
fails.  It leaves the user registered, but not logged in.  The solution for
SPAM registrations is to use confirmation.

This fix also adds a "catch" for a save failure.  If the user topic
cannot be saved (for example due to AntiWikiSpamPlugin), the fail is
caught and the registration is backed out.

Finally the removeUser function removed the user from the password
manager,  but did not remove the user from the WikiUsers mapping
topic.  Refactored the addUser code to also support removing a mapped
user.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@14198 0b4bb1d4-4e5a-0410-9cc4-b2b747904278