Item12023: Add filter by log record contents
For the simple File logger only, a regex can be added to the file
specification to log matching records to a unique file.

Also fixed a bug where the eachEventSince on FileRolling was not
processing the current file.

Fixed up unit tests.  Some of the verify_ tests should really be
specific test_ tests because they are specific to a single logger.
Improved skip logic to skip tests that don't make sense for the
LogDispatch logger.

Found some verify_ tests that were not using the test fixture.

git-svn-id: http://svn.foswiki.org/trunk@15371 0b4bb1d4-4e5a-0410-9cc4-b2b747904278