Item11571: Still broken unit tests
This is closer,  but the tests are still broken.  The search $summary
returns literal macros like %TWISTY and %INCLUDE.  If I write them into
the "Expected results" section, then they expand.  If I escape them with
!%TWISTY then the % is encoded as &#37;

Not sure if this needs a fix to the compare code, or more clever
encoding of the expected results.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@15425 0b4bb1d4-4e5a-0410-9cc4-b2b747904278