Item11746: Change email instructions are wrong
When the Password Manager is managing emails, the address in the Users
topic is never used.

This change still isn't 100%.  It decides if the email is in the topic
depending on whether or not the password file is writable.   It really
ought to be asking if the password manager is managing emails.

If password manager is managing emails
   if not writable,  email change is temporarily disabled.
   otherwise the change form should be active.

Also this topic is very specific to the TopicUserMapper.  However the
change email address API could be used by any mapper.  This topic should
probably be moved to System, and only include the topic user mapper
information when it's enabled..

git-svn-id: http://svn.foswiki.org/trunk@15537 0b4bb1d4-4e5a-0410-9cc4-b2b747904278