Item12225: Patch files matching by MD5 sig.
The patch should be able to apply to multiple releases.  However since
Text::Patch doesn't do fuzzy patching, and we want this to be relatively
safe, verify the md5sum of each file before attempting to patch it.

Still need backup of the files to be patched, and some documentation
might be helpful.

git-svn-id: http://svn.foswiki.org/trunk@15960 0b4bb1d4-4e5a-0410-9cc4-b2b747904278