Item11267: Add PatchFoswikContrib to release
Adding this to the branch doesn't actually do anything other than
satisfy a dependency that would be triggered when installing extensions
using the new Version string.

By addding the extension to 1.1.6, it prevents the installation of the
patches, since they are actually installed by the POSTINSTALL exit.

This also eliminates dependency errors reported by pseudo-install.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@16135 0b4bb1d4-4e5a-0410-9cc4-b2b747904278