Item12353: Add checker for MaxLSCBackups
If the lib directory is write protected, which is the default
configuration for the RPM and Debian packages, then the user
is stuck unable to save the configuration because configure cannot
create the backup.

Adding a checker to disable backups if the directory is not writable.

Also moving {MaxLSCBackups} into the "General Path Settings" tab, so that the
checker runs during an initial configuration as well.   The scenario
here is that if the admin knows lib is write protected and "touches"
LocalSite.cfg to make sure a file exists for configure to write, then
the save was failing again because of the default backups.

git-svn-id: http://svn.foswiki.org/trunk@16405 0b4bb1d4-4e5a-0410-9cc4-b2b747904278