Item12371: skip empty key/value pairs when caching requests
Additionally, only treat a line as a request param terminator if it is a
single "=" with nothing after it. This theoretically allows people to set
values on an empty field, though that doesn't conform to the usual standards
and I really can't conceive of anything that would legitimately do it.


git-svn-id: http://svn.foswiki.org/trunk@16482 0b4bb1d4-4e5a-0410-9cc4-b2b747904278