Item12616: Fixes for perl 5.18
The HTTP headers can now be generated in any order.  The unit tests were
sensitive to location of the Status: line.  Change the tests to allow it
on any line in the headers.

This also exposed a bug in Render.pm which will have caused inconsistent
escaping of characters in the HTML.  The %ESCAPED hash defined the
backslash as
      '\\' => '%5C',
This actually inserts a single backslash.   But in the regex, a single
backslash is an escape, which caused the backslash to be ignored.

With the changes in hash key order, the backslash occasionally ended up
as the final character in the regex, which escaped the closing square
bracket, breaking the regex.

    my $chars = join( '', keys(%ESCAPED) );
    $txt =~ s/([$chars])/$ESCAPED{$1}/g;

Changing the hash key to '\\\\' inserts an escaped backslash \\ as the
actual key, which causes it to be matched as a backslash in the regex.

git-svn-id: http://svn.foswiki.org/branches/Release01x01@16985 0b4bb1d4-4e5a-0410-9cc4-b2b747904278