Item12360, Item12023: LogDispatch updates
Update documentation to reflect changes
Update Config.spec with catch-all token, and fix typos.
Update all the default config with the catch-all tokens.

Separate the field hash from the LogDispatch parameter hash.  There was
a possibility that internal Log::Dispatch keys might conflict with
logged fields.

Rename any foswiki specific fields in the parameter hash to include
_underscore prefix. Help avoid future collisions.

Update the _flattenLog routine to append any fields omitted from the log
layout into the "*" log token.  Better coverage if additional fields are
added to the log calls.

Update unit tests for the catch-all token

Add a unit test for the Filtered style events.

SMELL:  The eachEventSince is documented as ignoring filtered files
however that's not implemented, as the unit test shows.  It would be
better to have some way to specify the file.   Currently it depends up
on hash order if more than one file covers a level.

git-svn-id: http://svn.foswiki.org/trunk@17212 0b4bb1d4-4e5a-0410-9cc4-b2b747904278