Item12104: Cleanup some code
The "numLevels" variable wasn't used and doesn't need to be in the API.

A couple of class files had the wrong package names, and Obfuscating.pm
redefined a method for no reason, removed unneeded code.

The API doc's needed some updates.   API version wasn't documented, and
the plan to allow passing a field list will not be implemented.

git-svn-id: http://svn.foswiki.org/trunk@17225 0b4bb1d4-4e5a-0410-9cc4-b2b747904278