Item11267: Sync up with svn exits
pre-commit now checks the TOPICINFO lines.

Changed the logic for checking perl files.  We were only checking .pl
and .pm files.   Instead, bypass files with known non-perl suffixes. But
any other files,  check for the shebang if it's not a .pl / .pm file.

Warn but allow the commit to go through if a rcs,v file is checked in.

Don't do any checking for the CompareRevisionsdAddOn sample file.  It
needs to be a specific rev & version for the demonstration to work.

git-svn-id: http://svn.foswiki.org/trunk@17429 0b4bb1d4-4e5a-0410-9cc4-b2b747904278