Item12839: Update documentation, add remove description
Update the Func and EmptyPlugin documentation to explain the changes in
the REST handler security options.

Also put back the 'description' column in the %RESTHANDLERS% macro. I'm
thinking that for users not in the AdminGroup, we can redact all of the
security information and only show the description.

http://foswiki.org/Development/GuidelinesForSecureExtensions has been
updated with documentation as well.

git-svn-id: http://svn.foswiki.org/trunk@17515 0b4bb1d4-4e5a-0410-9cc4-b2b747904278