Item12934: Make email autoconfigure a bit more forgiving
It needs to better identify the dependencies on IO::Socket::SSL, and
Net::SSLeay  (the openssl bindings).

If the dependencies are not installed, just try plain old SMTP on Port
25.  That still might work for windows users rather than falling back to
the unix mailprogram

Disable $Foswiki::cfg{Email}{SSLVerifyServer} by default.   yes
SSL certificate validation is good to prevent server spoofing and other
threats.  But configuration is complex, and adds dependencies on
providing the list of root Certificate authorities that you trust to
issue certificates.

git-svn-id: http://svn.foswiki.org/trunk@17746 0b4bb1d4-4e5a-0410-9cc4-b2b747904278