Item10764: Item12027: Log using utf8
The logger will now always write in utf8, encoding into utf8 as
required.  This resolves some issues caused if character encodings
unsupported by the site character set end up in a logger call.

It also avoids switching encodings on the fly depending on the
characters in the log call.

Also move around the monkey patching of time in the unit tests to avoid
an infinite loop in the test if a test fails without undoing the patch.

git-svn-id: http://svn.foswiki.org/trunk@17772 0b4bb1d4-4e5a-0410-9cc4-b2b747904278