Item11267:  Delete unneeded ImageGalleryPlugin
This plugin is not part of core, and was incorrectly included into the
distro repository.  It can be removed,  but the procedure is complex, so
moving it out of the way for now.  ImageGalleryPlugin should be
separately cloned. from https://github.com/foswiki/ImageGalleryPlugin

The command to completely remove it from the index is:

git filter-branch --index-filter 'git ls-files -- ImageGalleryPlugin |
xargs git update-index --remove' --tag-name-filter cat --prune-empty -f
-- --all;

Note that this will rewrite the commit hashes of all commits and will
break the HEAD pointer for anyone who has cloned the repository.  So
when we eventually want to get rid of it, the procedure will be:

0) Clone the repo as a backup
1) Run the filter-branch
2) git push --force all

Each developer must then:

3) git fetch   (NOT PULL)
4) for each local branch: git rebase --onto origin/master  master <local branch>
5) git checkout master
6) git pull --force

See
http://stackoverflow.com/questions/7561998/rebasing-everyone-onto-changed-git-history-after-filter-branch