Item11267: Use better default for git url
The git:// protocol isn't advertised by github, and is read-only, can't
be used for push / update operations.   So switch default to https://

Also, if the current repo is a SSH based (git@github.com) repo, then
override the git: / https:   to be a SSH connection.  This way a
developer with ssh keys set up can push without password prompts.