Item12952: Changes to configure security.
Defines a new configuration setting: ConfigureFilter.  If configured,
this takes precedence over the isAnAdmin test, otherwise all users in
the AdminGroup can access the configuration.

Add a log call so we know who is using configure.

Move the old configure password under Security And Authentication tab as
an expert parameter.  Still need to figure out what to do with this.

(Part 1.  Also needs changes to ConfigurePlugin)