Item12952: Finish bootstrap of view w/ short urls
The "installation instructions" to go along with this.
  - User installs foswiki
  - Configure apache with ApacheConfigGenerator
  - Visit the anticipated home url:
 ex: http://mysite.com/foswiki/Main/WebHome

  - The Bootstrap warning has link to configure.  Click configure and
    the view url is correctly bootstrapped.

I had to stash the initial view URL into a session variable because the
jsonrpc scripts need it, and they are invoked with POST.

The session variable code ... I think that could be moved to
ConfigurePlugin or the UI::Configure -  wasn't quite sure where to put
it,  but it works okay in Foswiki.pm

Also this changes the NOT_SET array to support multi-level parameters in
the bootstrap.