Item12952: Improve behavior of bootstrap mode.
Make admin decisions based upon isBOOTSTRAP as early in execution as
possible, so that later changes to isBOOTSTRAP are not recognized.

By setting admin user early, it "sticks" when the initial foswiki
session is created.  This leaves the user logged in as admin after the
initial configuration is saved, so that they have a chance to make
changes to AdminGroup without getting locked out.