Item12952: Restore files deleted from Configure
Util.pm is used extensively in Extensions installers.

The majority of the ConfigureTests actually test the Package installers,
shebang rewriting,  file mapping functions, etc. required to install an
extension into a customized foswiki

Next commit will rename it from ConfigureTests to
ExtensionInstallerTests

The tests are broken because the output of the Dependency check has been
reformatted to be pretty in the new configure,  so it breaks the
dependency reports printed when installing extensions.