Item12952: Don't scan INC for checkers
We know the exact path, so just require the module, and test the
results. This knocks 1/2 second or more off of the each rpc
request that loads the checkers.  Calling FileUtil::findPackages only
makes sense when wildcard searching for a list.

(Running under NYTProfile it reduced a call jsonrpc from 13 seconds to
7.  )