Item12952: Finish removing NOT SET handling.
Also fix up a couple of checkers that were modifying the configuration:

 - ../ConfigurePlugin/lib/Foswiki/Plugins/ConfigurePlugin.pm

Removed NOT SET processing.  It's no longer used.

 - lib/Foswiki/Configure/Checker.pm

Eliminated guessMajorDir and guessDirectory.  These are not used by
configure. Checkers don't set values

 - lib/Foswiki/Configure/Checkers/BasicSanity.pm
 - lib/Foswiki/Configure/Checkers/DefaultUrlHost.pm
 - lib/Foswiki/Configure/Checkers/SafeEnvPath.pm
 - lib/Foswiki/Configure/Checkers/ScriptUrlPath.pm

Clean up NOT SET processing and guessing of values.

 - lib/Foswiki/Configure/Checkers/DispScriptUrlPath.pm.

Removed.  Once the NOT SET and guess processing was removed there was
nothing left. This variable did not appear to be used.