Item12952: Improve extension installer
Change "Install" button to "Study".  And add some documentation / page
headings to make it more readable.

Add a warning about the automatic installation of deps if there are
missing deps.

Add "Simulate" and "Install without Dependencies" buttons to the
Dependency report.

Note:   The  extension installer is not working. There is something
going on with the repository parameter.  That's unrelated to this
checkin.