Item12952: Hack calling inconsistency
Probably better fixed in the caller, so that two different conventions
are used.   But this works.

Also, don't update the configuration if Simulate option is requested.

Note,  There are several places where encapsulation of package is broken
by direct references to ->{_options} is used.  These should probably be
changed to ->option() calls.