Item12952: Add access control to Dependency Report
Display the location where each module was found in the @INC path.  This
should help debug issues when modules are provided by system, local
cpan, and/or Foswiki.

Also add access control to the DependencyReport.  Since this exposes
perl module versions and now, file system locations,  it really should
be restricted to users with access to Configure.

SMELL:  This is the 3rd place that evaluates the {ConfigureFilter}.  It
probably ought to be refactored somewhere, but I'm not sure where to put
the code.

It is intentionally not part of topic ACL rules.  We don't want a wiki
admin to be able to promote themselves to configure access.