Item12952: PERL types get corrupted when updated
Fixes two issues:

1) When a multi-line set is submitted through jsonrpc, the
untaint regex is not multi-line and truncates it at the first line.

2) Type PERL are stored as "ref" types HASH or ARRAY,  but are input as quoted
strings when saved from jsonrpc set.

SMELL ...   Simple quoted strings of type PERL still get corrupted.  After a save,
the quotes have been stripped.

  - $Foswiki::cfg{Plugins}{ConfigurePlugin}{Test}{PERL} = 'PERL';

ends up stored as
  - $Foswiki::cfg{Plugins}{ConfigurePlugin}{Test}{PERL} = PERL;

This results in a corrupted LocalSite.cfg.  This can be fixed by
starting with the following in the Spec file:

  - $Foswiki::cfg{Plugins}{ConfigurePlugin}{Test}{PERL} = '\'PERL\';';