Item13063: Improve logging of changes
Don't truncate the field from the logs,  we only need to truncate with
ellipsis for the UI, so the screen remains readable.

Also, encode vertical bars before generating the TML table, so that it
isn't corrupted.

Crawford,  this still is not completely correct. I eliminated your
conditional use of $reporter.   PERL fields were not having their
changes reported to the UI until I made $reporter unconditional.