Item13079: Item13081: Better bootstrap for tools/configure
Change configure to make use of the Load::bootstrapConfig() function.
This will apply consistent guesses in shell and web environments.
And use the guesses from bootstrapConfig as the defaults in the prompts.
That cleans up the ../ relative paths.

Add an "optional" flag to the _prompt routine, so that settings
permitting empty or undefined can be configured.

There are still a few things that should be done.  It would be good if
Load::bootstrapConfig() could use the $reporter instead of returning
strings.

Also if for some reason bootstrap is unable to guess a critical
directory, it will still die.  That needs to be handled better.

Fixed up the docs for -set.  Maybe it's a GetOpt issue, but I can't get
-set {key} 'value' to work.   It has to be -set {key}='value', the equal
sign appears to be required.