Item13068: Bootstrap the {Store} related settings
Add a bootstrapStore method in the Store contrib modules.  This method
examines the system or file systems to bootstrap that Store's settings,
and determine if the store is applicable.

add _bootstrapStore to Configure::Load.  It invokes each Store contrib's
bootstrap method if available.  It also sets the core managed Store
settings:  {Store}{SearchAlgorithm} and {_grepProgram}

Comment out the settings in Foswiki.spec, and add them to the prompted
settings in tools/configure