Item13079: Bootstrap doesn't gain admin rights.
If the user has a valid Foswiki session cookie, and the
working/tmp/cgisess_ file exists, then the bootstrap process fails to
set admin user, and runs as the current (stale) session user.

When loading the user from the session cookie, detect that we are in
bootstrap mode, and upgrade the session to admin.