Item13023: Unit tests are not happy with fork to grep
The TestBootstrapPlugin loops in the fork to `grep -V` during test to
determine if there is a valid grep program on the default path.

SMELL:  I was unable to debug this.  Perl debugger also failed due to
the fork. Not sure what is going on here.  Temporarily skipping the
probe.  I'd rather use File::Which to actually find the path location
rather than trusting that the path will be stable.  But that's not a
core module, and it's not worth adding another dependency.