Item13142: Don't trust cache when saving file
Reread the file before actually updating, make sure that the cache can't
have been corrupted somewhere along the way.

Make the try/catch clauses consistent.  Not sure that this is 100%
correct, but we are missing something somewhere.

Note:  This code is running live on foswiki.org,  with additional checks
enabled to refuse to write a truncated htpasswd file.