Item13167: deprecate CGI for rendering purposes
This removes some of the most bogus calls replacing them with straignt string operations.
The rest could be removed as well but needs more testing, e.g. wrt escaping HTML attribute values.

There is a minor speedup gain part of not calling CGI::foobar(). No.1 issue is to lower the
dependency on CGI.pm, i.e. when there really is no point in using it.