Item12381: Clarify email wizard documentation.
Hide fields unless they are applicable.  Ie don't show the configure button until
at least the webmaster email is supplied.  Don't show user/pass if the server
name is not supplied.   Make it clear that if the server is able to send email,
then it just might work without smtp configuration.

Move the Send a test email next to the Auto-configure button.  so the actions
are below the configuration settings.