Item13220: An experiment in email flow
Crawford, please revert if you don't like,  easier to do it this way
than attaching patches.   Trying another flow for the email settings.

 - Move EnableEmail onto the Advanced settings panel.  The Wizard will
   control it.
 - Move Auto-config and Test email buttons up to WebMasterEmail.
 - Duplicate the Auto-config button onto the SMTP SERVER field
 - Add a checker for WebMasterEmail, and Remove the EnableEmail checker.
   The WebMasterEmail checker subclasses the EMAILADDRESS checker and
   provides more meaningful errors than the generic checker for a
   missing email address.

I think that this gets the auto-config flow onto the first page and
minimizes the chance of the user enabling email with an invalid
configuration.

One issue.  When the Wizard sets EnableEmail, it fails to trigger the
re-check of WebMasterEmail, so there is a stuck warning.