Item13278: Ensure regex defaults agree with cfg values
This fixes configure, but there are some operational issues.

The NameFilter regex is passed verbatim to the NAMEFILTER variable in
the DOM.  core/pub/System/JavascriptFiles/foswikiWebTopicCreator_src.js
then uses it to filter topic names.

However the {NameFilter} regex saved by configure is incompatible with
javascript, so the filter fails.