Item13278: Store regexes as quoted strings.
This is consistent with Foswiki 1.1, which stores regexes as simple
quoted strings.

Also fixed a bug that embedded single quotes in strings will break in
types.js because the non-greedy regex stops at the first embedded quote
instead of the delmiter.

And a typo in the name of the Reporter.stripStacktrace function.