Item13178: partial fix for expanding variables in urls.
TinyMCE expands some macros in the left side of [[wiki][links]]
and WysiwygPlugin will convert hard coded URLs back to macros.

These lists ought to match, so that anything that TinyMCE expands will
be reversed by Wysiwyg during save.  This works for prefixes,  PUBURL,
ATTACHURL, etc,   but not for embedded web / topic macros.

This fix adds exceptions for SYSTEMWEB, WEB and TOPIC.

This is a *parital fix* because there are still issues with %SCRIPTURL%.
An example  [[%SCRIPTURL%/configure][configure]] will be corrupted to
[[bin.configure][configure]].