Item13294: Item13297: Issues with UserRegistration
The documentation for the REGISTRATIONGROUPS feature didn't note that
the extraField_groups needed to be enabled in the user registration
screen.

The MAKETEXT strings for the REGISTRATIONGROUPS feature couldn't be
translated because of the way it merged in the singular / plural groups.

Documentation for UserRegistration didn't document the new registration
form layout using the FORMAT macro.