Item13302: changing a form breaks char encoding
At some point CGI started to treat form values differently, i.e.
when they are byte strings as we have it inside Foswiki holding
unicode chars.

This is at least a problem using CGI-4.13. It recodes them using
html entities, but fails to guess the char encoding thus freaking
the strings up.

Older CGIs < 4.00 are not affecting Foswiki.

The work-around is to avoild the HTML generators in CGI as far
as possible and just do plain string operations.