Item13294: Address inconsistencies in terminology
LoginName and UserName needed some reconciliation.

"username" is a thing,  so it's spelled that way, No bumpy text, since
it's typically not entered that way.

"WikiName" is also a thing, but is spelled with bumpy text, since it
noramlly requires  it for linking.

The term "LoginName, log in name,  login name, etc.  were used very
inconsistently and are retired.   Where clarity is needed, it's written
as "Login username", but most often username.

The LoginName topic has been renamed to UserName and moved from the
TopicUserMappingContrib to the core.  It is MAKETEXTed, and renders with
conditional text.  If AllowLoginName is not enabled,  it should be clear
that the UserName is the WikiName.     If AllowLoginName is enabled, it
provides details on valid usernames.   The WikiUsers related text is
also conditional and will not display if there is no WikiUsers topic.