Item13185: Rework the URL path parser
Do better validations of the web path components and topic names.

Preserve the invalid names, so that they can be reported in oops
messages.

Add oops messages to Preview, Rename, Save and View when they are called
with an invalid web or topic name.  Previously they were often
sanitized to a valid default web or topic name.

Add an error message for invalid characters in a topic name.  Revise the
bad_web error to show the incorrect value.

Client tests were a test error.  It was setting a query parameter into
the query_path.

ViewScriptTests have 3 changes:
 - /Sandbox/WebIndex/?topic=WebChanges   Previously the WebIndex would
   be preserved in the web path,  Now it's detected as a valid topic
   name and overridden by the topic= parameter.
 - /Someweb/Subweb?topic=WebChanges   Prevously the Subweb was ignored
   because of no trailing slash,  and overridden by the topic name,
   returning /Someweb/WebChanges.   If the Subweb is a valid webname, it
   now returns  /Someweb/Subweb/WebChanges.
 - /A:B/WebHome  used to return an empty webName.  It now sets the
   default UsersWeb webname, but also sets invalidWeb,  so that an error
   will be thrown.