Item9693: Restore WikiSyntaxSummary
It's translated, so good for non-English users.

It's INCLUDED for the old wikitext editor, which isn't used when natedit
is enabled.  But when not enabled, it was generating INCLUDE errors
and no help text.   The other help topics are much too big.  This is a
reasonably summary.

Rerun xgettext to merge in the strings.  Most were already translated,
so shouldn't be too painful.