Item13322: Add support for an "onSave" checker
This extends the Spec definition with "ONSAVE".  If set, the Save wizard
will call <checker>->onSave( $reporter, $key, $value ) for the key
before the value is processed for save.

The handler can just do cleanup work, such as in this case, removing the
language.cache file, so that the cache is refreshed. It can also modify
the value.  It should not change any other configuration values.

Modification is intended for custom encoding, such as hashing a
password.