Item13315: Address usability of the admin password
Move the password hashing code from tools/configure into the
onSave handler of the password checker.

This allows the password to be set in plain text through both the CLI
tools/configure as well as with web UI.  It detects if someone has
pasted in an already hashed password, and will not double-encode it.