Item13302: reverted the utf8 workarounds
CGI-4.11, 4.12, 4.13 are buggy and not usable in an utf8 setting on Foswiki
This has been reported to the maintainers. CGI-4.14 will fix the issue so that
we won't have to deal with it anymore. As Jan told me the fix wasn't working
correctly for all unicode code points anyway thus still escaping a lot more than
we want.