Item13391: Major refactor of extender.pl and tools/configure
tools/configure has some very big advantages when used to install
extensions.
   - It merges in the config.spec
   - It allows the extension to be enabled
   - It saves the configuration

The old extender.pl can't do any of that.

This commit changes extender.pl to merely reconfigure the commandline
arguments and "do" tools/configure to install the extensions.  This
greatly simplifies the code, and centralizes all installation work into
the Wizard and Package.pm.

In order to make tools/configure more "friendly" to the cli user, this
changes the $reporter to use print STDERR statements instead of doing a
"Data::Dumper" of the reporter messages hash.