Item13456: Support utf-8 in admin password
This also prevents a crash if someone attempts a utf8 password,
regardless of the actual password.  The hash routines all require
bytes, not unicode characters.

Also document a possible change to tools/configure so that utf8
can be used for entries from the shell.  Note that utf8 is still usable
if tools/configure is run with the -CA perl option.

Finally, add a "use utf8" to LocalSite.cfg.  This isn't really needed
but would accommodate any user who manualy edits a utf8 name into the
config, such as the Webnames,  email addresses, etc.

Update documentation to explain how to get utf-8 characters into the
configuration.