Item13385: Allow \| vbar escapes in form values
Forms are documented as supporting backslash in the initial values
definition.  The table parser does not support \|,  and a simple
escape of \| breaks things.

Since the Foswiki::Form uses the generic Tables::Parser, this
change detects when parsing a form,  and handles the \| escaped vbars
only in the initial values column.