Item13599: Don't skip files if no session
Extension installer tries to check-in topics that have history in the
store, or are flagged to checkin in the manifest.  When the installer
is run from the CLI though, there is no session and checkin won't work,
so it skips the file.

Just copy the file if no session available.