Item13560: Add web interface for plugin settings merge
This doesn't completely work.  There appears to be a bug in the configure.js
javascript. The {Module} and {Enabled} settings don't "stick"
when set by a wizard.  Any other settings are working fine.

Because the Config.spec files are actually loaded and merged in
=Wizards::Save=, there needs to be some way to force a save of the
configuration even when no settings are actually changed.  This change
adds a require_save() function to $reporter.  The Javascript can then
use that to enable the Save button even when no changes are in the DOM.

None of this is an issue from the CLI.  Running the merge wizard from
the shell just works.