Item13619: Remove RevComment from templates.
This removed the RevComment customization from the core templates. The
function will be added by a skin override shipped with RevCommentPlugin

This is safe to remove in a core patch because the RevCommentPlugin is not
currently functional on Foswiki 2.0.

Note that the RevCommentPlugin CSS is still contained in the
PatternSkin CSS files.