Item13917: Fix NameFilter and default values
Two issues.   The NameFilter and AttachmentNameFilters were using
excessive \\ escaping of characters.  The only things that need to be
escaped are regex meta characters.

The types.js code that stripped out escapes of embedded single quotes
was mathching on double-backslash.

There is still something not quite right.  The single-quote has to be
escaped in the perl string,  but does not need to be escaped in the
actual regex.

And the backslash really is required. I think it's working because the
escaped single quote really isn't an escape in the regex, but is an
escape for the stringification. So the backslash is magic. So the perl
code isn't correctly setting and clearing the escaped single quote.