Item13897: Moo-fication and UnitTestContrib adaptation
AddressTests is now passing all but one test. test_roundtrip fails.

- Most critical issue is existance of both method expect_failure in
FoswikiTestCase and attribute of the same name in Unit::TestCase. Due to
the nature of the method it is impossible to only request the value through
it as Moo/Moose require. There are only two solutions: either changing the
reader name for the attribute to something like get_expect_failure, or
renaming the attribute.  In both cases it still necessary to change the way
Unit::TestRunner accesses it. I decided that renaming would work better.
Now we have expecting_failure attribute and expect_failure method.

As there is no need to change any test suite this change could be
considered very much backward-compatible.

- Still, expecting_failure acts similar to the old expect_failure to
maintain compatibility throughout the code: it can be set with a reason
text which will be added to the annotations attribute; or with 0/undef
which will mean no reason to annotate. But in either case expecting_failure
will be set to true. Resetting is possible by a call to
clear_expecting_failure method.

- Unit::TestRunner now captures __DIE__ and rethrows it with
Foswiki::Exception on per-test basis.

- SMELL Not sure about passing $tester->annotations to set_annotations.

-- Foswiki::Address got to be a tricky one. :

- isA and type attributes are set to be private due to specific behavior of
their respective methods.

- _parse is now pure static method and doesn't mangle with the object.

- _atomizeXXX, _existScore are now pure static too.

- attributes with impact on object validity are now calling isValid()
method when changed. Should have no or minimal impact on performance.

-- Foswiki::Object

- added type checker method isaARRAY to be used in isa property of an
attribute.