Item13897: Fixed a bug affecting test_roundtrip.
It seems that whenever there is an attachment defined in resourse
Foswiki::Address.pm requires that the attachment has a name. The name is
expected to be in the second element of tompath array (tompath->[1]). But
test data for test_roundtrip containted a line defining an arrayref with
single 'attachment' string element which is not considered valid. Not when
DEBUG is set true at least. That line has been removed, AddressTests now
succeeds completely.