Item13897: Tests up to and including FormattingTests
ExtensionInstallerTests needed some special care from FileUtil.pm because
it didn't play well with BSD tar which sends it's output to stderr while
create a new archive and verbose operation requested. FileUtil now tries to
find GNU tar first and use it and falls back to BSD tar if fails. In latter
case it redirects stderr to stdout.

Fixed use of $@ instead of required $? for checking exit code of a
subprocess.