Item13897: Intermidiate commit
Many parts are to be Moo-fied before tests would run again.

Most notable change is that Foswiki::Form object has now be created by
means of method load() instead of new(). This is because generally it's not
a very good idea to overload Moo's new() and there is no other way to
bypass the object construction and fetch a form from the cache if it's been
loaded beforehand. Generally it's makes sense as previously
Foswiki::Form->new() call was not necessarily creating a new object after
all. Whereas load() does load a form either from disk or from the cache.