Item13917: Item13937: Restore backslash into NameFilter
If a backslash ends up in a topic name, it crashes search.  See
Item13937 for the search crash.

Also need to process the backslash escaping in configure types.js, so
that the backslashes don't get doubled again.

Also adds forward slash to the topic name filter. A forward slash is a
web delimiter, and should not be in the topic name.  And also filter it
from attachment names, as it's a directory path separator in the file
system.

And this also removes semicolon from the attachment name filter.