Item13897: Another huge commit, tests are passing up to and includind HoistREsTests
- Mostly borning routine convertion to Moo. Processed Render, UI, and Logger
subsystems and some of Foswiki::* classes.

- Fixed the method of checking for RCS co utility in
FoswikiStoreTestCase.pm. The previous method was unreliable and was causing
STDOUT redirection to /dev/null in my case. Now $PATH is manually traversed
and checked agains co executable. Reccommended for porting back to the
master.

- Getting rid of @_newParameters meta-protocol. Avoiding its use in newly
converted classes.

- Getting rid of finish() method whenever possible. DEMOLISH and
auto-destruction is the target.

- Added methods transmute() and rethrowAs() to Foswiki::Exception.
transmute() coerces any passed object into predefined class. rethrowAs uses
this functionality to always rethrow an execption as a specific class. It
differs from rethrow() in the way that latter attempts to preserve the
original exception it rethrows.

- Bugfixes, of course.