Item13897: Bug fixes and few more tests.
- Commented out the code for $Foswiki::cfg{Store}{ImplementationClasses}
support. It has to be reimplemented in a completely different way.
Preferable as a part of new plugins subsytem.

- Fixed Foswiki::load_package() to check if module exists in the perl
symbols table. This is the only right way if a module being loaded is
already inlined into the code.

- RobustnessTests: with utf8 on when 'use locale' is mixed with 'use
feature "unicode_strings"' symbols \x85 and \xa0 become additional white
spaces (next line and &nbsp; correspondingly). As this combination of
pragmas is the default for this branch added these symbols to the expected
value in test_filterTopicName.

- Pulled the latest filter regexps and test expects from the master branch.