Item13897: All top level tests are passing!
Well, all but Selenium and SemiAutomatic tests. Left them for later.

- Converted Tables code.

- Added named parameters to load_package. Currently the only key it accepts
is `method` which defines what sub to check in the module to determine if
it was loaded before. This is to fix situations when package is by
occassion is refernced before being actually loaded from a file. In this
case the symbol table would contain a reference to the module. By testing
if the package can(method) we make sure that it is already loaded normally.

- Added method load_class which is a wrapper for load_package( $package,
method => 'new' ).

- Few bug fixes.