Item9790: Item2266: Add password reset form and summary
Hide the detailed registration report. Only include a table of the
registerd users.

Generate a bulk password reset form for any bulk registered users that
didn't have a password set in the input.

Also only show the registration report with the timestamp on the report
is newer than the timestamp on the input.