Item13883: Update docs missed from Foswiki 2.0
Added docs on the use ALLOW*HISTORY and ALLOW*RAW settings.

Added config reference on where to find URL controls in INCLUDE

Added example settings to the WebPreferences topics for System, Main,
Sandbox and _default webs.

Also, add configure to the list of protected scripts by default.  It has
good internal security, but it still should be on the list of scripts
that require authentication.

Also added jsonrpc to the list of scripts requiring protection for sites
that are completely closed.  jsonrpc should normally not be restricted.