Item14033: Restructure engine initalization
Thanks to vrurg for the suggestion on how to restructure
Foswiki::Engine.

The creation of the Foswiki::Request object needs to be delayed until
the "action" has been extracted from the path. This allows a custom
request object to be created to handle special script path parsing
requirements.

EngineTests needed a more modern switchboard.

Also, this fixes a typo in the request parser code.  And add back in
the urlDecode missed from the original parser.