Item13897: Saving intermidiate changes.
Notable changes:

- Started convertion of engine into a information center for the startup
stage. The preparePath() method is the first victim: it now returns a
hashref of action/path_info/uri keys.

- Foswiki::App::handleRequest is incorporating Foswiki::UI::handleRequest
functionality.

- New exception Foswiki::Exception::HTTPResponse. Used to send a HTTP
response to user. Not an error/fatal exception though might be used to send
error HTTP statuses to the user. By default it is using
$Foswiki::app->response object including the status code.

- Foswiki::Exception::Engine now inherits from
Foswiki::Exception::HTTPResponse.

- Foswiki::Request is incorporating code from Item14033 branch by George
Clark.

- Foswiki::Request is now self-initializing using $Foswiki::epp->engine. It
shall be kept in mind that whatever data comes from the engine it may be
later changed/overriden in the request object. Therefore they're not
interchangable. Actually request should be the only user of the engine
unless there is The Really Good Reason to use itby some other subsystem.