Item14033: Implement Request::Rest
Also fix a bug in the Request::parse function.  Need to use
the path's leading slash as a hint that the first part will
always be a web.

Also add some further checking to the UI::Rest function.

Found an issue with SubscribePlugin.  It passes things like wild
card * characters in the topic query param.  For now, the topic
name validation in Rest handler has to be bypassed.