Item13897: Few more touches on engines and configuration.
Changed the way we detect Engine to be chosen when not defined by
the configuration Engine key (mostly while bootstrapping). Lets think of
engine as a driver. We simply try all engines listed in EngineList
configuration key and call their static method probe(). Within this model
CLI considered the last resort option; its probe() always return true and
it has to be the last in the list.

Uploads handling still needs retouching.

- Split prepareHeaders engine method into _prepareHeaders and _prepareUser.
Added user and headers attributes to the Engine class.

- Removed prepareCookies from the Engine. Cookies are now pure request
property and are been set using 'Cookie' header.

- Request now initializes headers using engine's data.

- Adapted JQuery part of Subscriber and Twisty plugins.